chore: move client cache into a "private" module
