+++ test/functional/keyrings/aws_kms/test_client_cache.py
@@ -5,7 +5,7 @@ import pytest
 from botocore.config import Config
 from botocore.session import Session
 
-from aws_encryption_sdk.keyrings.aws_kms.client_cache import ClientCache
+from aws_encryption_sdk.keyrings.aws_kms._client_cache import ClientCache
 
 pytestmark = [pytest.mark.functional, pytest.mark.local]
